aasd

+++ test_gleu.py
@@ -111,7 +111,7 @@ def custom_collate_fn(batches):
     tokenized_inputs["labels"] = torch.LongTensor(labels)
     
     if task == "stsb":
-        tokenized_inputs["labels"] = torch.FloatTensor(labels)
+        tokenized_inputs["labels"] = torch.LongTensor(labels)
         
     
     return tokenized_inputs
